Changed to reflect new guidelines.

+++ trunk/TAO/examples/Simple/time-date/run_test.pl
@@ -9,15 +9,27 @@ use lib "../../../../bin";
 require ACEutils;
 require Process;
 
+$status = 0;
 $iorfile = "./ior";
 
-$SV = Process::Create ($EXEPREFIX."server$EXE_EXT");
+$SV = Process::Create ($EXEPREFIX."server$EXE_EXT", "");
 
-ACE::waitforfile ($iorfile);
+if (ACE::waitforfile_timed ($iorfile, 5) == -1) {
+  print STDERR "ERROR: timedout waiting for file <$iorfile>\n";
+  $SV->Kill (); $SV->TimedWait (1);
+  exit 1;
+}
 
-$status  = system ($EXEPREFIX."client$EXE_EXT -f $iorfile -ORBSkipServiceConfigOpen");
+$CL  = Process::Create ($EXEPREFIX."client$EXE_EXT",
+                            "-f $iorfile -ORBSkipServiceConfigOpen");
 
-$SV->Kill (); $SV->Wait ();
+if ($CL->TimedWait (60) == -1) {
+  print STDERR "ERROR: client timedout\n";
+  $status = 1;
+  $CL->Kill (); $CL->TimedWait (1);
+}
+
+$SV->Kill (); $SV->TimedWait (1);
 
 unlink $iorfile;
